Fix issues inviting submitter as a reviewer
[BPHH-1341]

- fixes issue where invitation token is not unique due to submitter
  getting updated before the inviter is destroyed. Instead of destroying
  only if submitter is updated succesfully, use a transaction so the
  destroy can happen first but will roll back if update is unsuccessful.
- fixes issue inviting a reviewer using the same submission email as an
  existing submitter. The unique email should only be enforced across
  reviewers/RMs to avoid duplicate invitations.

+++ app/services/merge_submitter_with_invited_user.rb
@@ -29,7 +29,12 @@ class MergeSubmitterWithInvitedUser
           invitations_count
         ],
       )
-    submitter.update!(invited_user_params)
-    invited_user.destroy!
+    submitter.assign_attributes(invited_user_params)
+    if submitter.valid?
+      ActiveRecord::Base.transaction do
+        invited_user.destroy!
+        submitter.save!
+      end
+    end
   end
 end
